runtime: Move out code from the lib.rs to the integration module
