+++ runtime/src/integration/mod.rs
@@ -2,5 +2,6 @@ pub mod content_working_group;
 pub mod forum;
 pub mod proposals;
 pub mod storage;
+pub mod transactions;
 pub mod versioned_store_permissions;
 pub mod working_group;
